requirement.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
pydantic==2.12.5
pydantic-core==2.12.0